Created factory module; user_factory moved, fixed and improved.
