+++ movienager/accounts/factory/user_factory.py
@@ -1,0 +1,21 @@
+import factory
+
+from accounts.models import User
+
+
+class UserFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = User
+        django_get_or_create = ("email", "username")
+
+    first_name = factory.Faker("first_name")
+    last_name = factory.Faker("last_name")
+    email = factory.Faker("email")
+    username = factory.Faker("user_name")
+    password = factory.Faker("password")
+
+    @classmethod
+    def _create(cls, model_class, *args, **kwargs):
+        """Override the default ``_create`` with our custom call."""
+        manager = cls._get_manager(model_class)
+        return manager.create_user(*args, **kwargs)
