--- movienager/accounts/management/commands/_user_factory.py
@@ -1,14 +1,0 @@
-import factory
-
-from accounts.models import User
-
-
-class UserFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = User
-
-    first_name = factory.Faker("first_name")
-    last_name = factory.Faker("last_name")
-    email = factory.Faker("email")
-    username = factory.Faker("word")
-    password = "fakepassword"
